basic ui to showcase time in game, with ability to change speed that time passes

+++ Assets/_Scripts/Simulation/ISimulation.cs
@@ -14,7 +14,7 @@ public interface ISimulation
     int GetSimulationSpeed();
 
     // Accessor methods for current time, day, month, year
-    int GetCurrentTime();
+    float GetCurrentTime();
     int GetCurrentDay();
     int GetCurrentMonth();
     int GetCurrentYear();
